+++ src/interface/mod.rs
@@ -1,9 +1,8 @@
-pub mod auth;
 pub mod cascade;
-pub mod guard;
 pub mod ic;
-//pub mod request;
-//pub mod response;
+pub mod request;
+pub mod response;
+pub mod state;
 
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
@@ -16,13 +15,8 @@ use thiserror::Error as ThisError;
 #[derive(CandidType, Debug, Serialize, Deserialize, ThisError)]
 pub enum InterfaceError {
     #[error(transparent)]
-    AuthError(#[from] auth::AuthError),
+    IcError(#[from] ic::IcError),
 
     #[error(transparent)]
-    CascadeError(#[from] cascade::CascadeError),
-
-    #[error(transparent)]
-    GuardError(#[from] guard::GuardError),
-    //  #[error(transparent)]
-    //  RequestError(#[from] request::RequestError),
+    RequestError(#[from] request::RequestError),
 }