+++ src/state/mod.rs
@@ -4,6 +4,17 @@ pub mod sharder;
 pub mod wasm;
 
 use crate::memory_manager;
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use thiserror::Error as ThisError;
+
+pub use {
+    core::{
+        app_state::AppStateError, canister_state::CanisterStateError, child_index::ChildIndexError,
+        subnet_index::SubnetIndexError,
+    },
+    wasm::WasmError,
+};
 
 //
 // MEMORY_MANAGER
@@ -17,5 +28,24 @@ const SUBNET_INDEX_MEMORY_ID: u8 = 2;
 const CANISTER_STATE_MEMORY_ID: u8 = 3;
 const CHILD_INDEX_MEMORY_ID: u8 = 4;
 
-//#[error(transparent)]
-//WasmError(#[from] wasm::WasmError),
+///
+/// StateError
+///
+
+#[derive(CandidType, Debug, Serialize, Deserialize, ThisError)]
+pub enum StateError {
+    #[error(transparent)]
+    AppStateError(#[from] AppStateError),
+
+    #[error(transparent)]
+    CanisterStateError(#[from] CanisterStateError),
+
+    #[error(transparent)]
+    ChildIndexError(#[from] ChildIndexError),
+
+    #[error(transparent)]
+    SubnetIndexError(#[from] SubnetIndexError),
+
+    #[error(transparent)]
+    WasmError(#[from] WasmError),
+}