+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.1]
 - moved loads of IC/canister-specific, and shared code from Dragginz into icu
+- have the old request/cascade/response code back and working
 
 ## [0.1.0]
 - ITS ALIVE!11!1!!